requirements.txt
Flask
slackclient
notion
firebase-admin
requests
sentry-sdk[flask]
